refactor(tools): split web-fetch helpers into src/tools_web.py
Move _PUBLIC_WEB_UA, _WEB_FETCH_TIMEOUT, _JINA_HEADER_MAX_LINES,
_JINA_LINK_RE, _HtmlTextExtractor, _NoRedirectHandler,
_validate_public_https_url, _filter_links, _extract_markdown_links,
_parse_jina_response, _read_body_capped, _jina_fetch, _raw_fetch from
src/tools.py into a new src/tools_web.py. Re-export all moved symbols
from src/tools for backward compatibility. Drop unused imports (ipaddress,
re, socket, HTMLParser) from src/tools.py. Update monkeypatch targets in
tests/test_tools.py for the moved helpers (src.tools_web.socket.getaddrinfo,
src.tools_web.urllib.request.urlopen/build_opener).
